+++ backend/tinyfarm/src/main/java/com/api/tinyfarm/repository/UserRepository.java
@@ -14,5 +14,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
     Optional<List<User>> findByLevel(Integer level);
     Optional<List<User>> findByHibernation(Boolean hibernation);
     Optional<List<User>> findByGender(User.Gender gender);
-    Optional<List<User>> findByEcus(Integer ecus);
+    Optional<List<User>> findByEcus(Float ecus);
 }
